arm64: kernel: FIQ support
Consider FIQs as if they were just IRQs.

Note that this switches FIQs from the invalid exception
handler to the IRQ handler. This shouldn't affect any
well-working system. It might be worth it to take a look
at this for patching though.

Instead of directly panicking, such systems will now
redirect the FIQ interrupt to the interrupt controller
driver.

As such, also make FIQs disabled when IRQs also are, switching
away from the model where the FIQs are just disabled when the
debug interrupts are disabled.

Signed-off-by: Stan Skowronek <stan@corellium.com>
Signed-off-by: Mohamed Mediouni <mohamed.mediouni@caramail.com>